修复未授权导致所有关卡都可以rate关卡的bug

+++ app/Services/Game/Level/CommentCommandService.php
@@ -60,6 +60,10 @@ class CommentCommandService
         RatingService $service
     ): string
     {
+        if (!optional($this->operator->permission)->can('COMMAND_RATE_LEVEL')) {
+            return 'Permission denied!';
+        }
+
         if ($this->level->rated) {
             if (Arr::hasAnyValue($this->options, ['d', 'delete'])) {
                 $this->level->rating->delete();
